chore: move sign up form-section into components folder
