+++ locales/ar/index.json
@@ -85,6 +85,7 @@
   },
   "sign up page": {
     "title": "إنشاء حساب جديد",
+    "full name": "اسمك الكام *",
     "nationality": "اختر جنسيتك *",
     "job": "قطاع الوظيفة *",
     "number": "رقم الهوية الوطنية *",
